containerapi: move to internal
